refactor(server): modify llm file struct
