+++ apps/server/api/src/llm/model/mod.rs
