+++ apps/server/api/src/llm/model/qwen3/quantized.rs
