+++ apps/server/api/src/llm/model/token_converter.rs
