+++ apps/server/api/src/llm/model/token_output_stream.rs
